+++ source/Base/Resources.h
@@ -2,7 +2,7 @@
 
 namespace Const
 {
-    std::string const ProgramVersion = "4.2.0";
+    std::string const ProgramVersion = "4.3.0";
 
     std::string const BasePath = "resources/";