+++ source/Gui/UploadSimulationDialog.h
@@ -9,10 +9,14 @@ class _UploadSimulationDialog : public _AlienDialog
 public:
     _UploadSimulationDialog(
         BrowserWindow const& browserWindow,
+        LoginDialog const& loginDialog,
         SimulationController const& simController,
         NetworkController const& networkController,
-        Viewport const& viewport);
+        Viewport const& viewport,
+        GenomeEditorWindow const& genomeEditorWindow);
     ~_UploadSimulationDialog();
+
+    void open(DataType dataType);
 
 private:
     void processIntern();
@@ -20,14 +24,18 @@ private:
 
     void onUpload();
 
-    BrowserWindow _browserWindow;
-    SimulationController _simController; 
-    Viewport _viewport;
-    NetworkController _networkController;
-
     std::string _simName;
     std::string _simDescription;
 
     std::string _origSimName;
     std::string _origSimDescription;
+
+    DataType _dataType = DataType_Simulation;
+
+    BrowserWindow _browserWindow;
+    LoginDialog _loginDialog;
+    SimulationController _simController;
+    Viewport _viewport;
+    NetworkController _networkController;
+    GenomeEditorWindow _genomeEditorWindow;
 };